==> var/00.20-pcm-sandbox.ipynb <==
%load_ext autoreload
%autoreload 2

from scipy.io import wavfile
import numpy as np

import sys
if "../src" not in sys.path:
    sys.path.append("../src")
from modist.modist_pcm_dataset import ModistPcmDataset, get_modist_pcm_collate_fn
from modist.mp3_info import get_lang_encoder

import json
with open("../data/mp3_split.json", "r") as fin:
    splits = json.load(fin)
train_mp3s = [x[0] for x in splits["train"]]
train_mp3s = [x.replace(".mp3", ".pcm") for x in train_mp3s]
pcm_list = ["../data/pcm/" + x for x in train_mp3s]

ds = ModistPcmDataset(pcm_list, speech_only=True)

len(ds.inventory)

ds[0]

def dummy_feat_extractor(x, *args, **kwargs):
    data = [xx.shape for xx in x]    
    return type('DummyFeats', (object,), {"input_values": data})

lang_encoder = get_lang_encoder()

from transformers import Wav2Vec2FeatureExtractor
feat_extractor = Wav2Vec2FeatureExtractor.from_pretrained("facebook/wav2vec2-base-960h")

collate_fn = get_modist_pcm_collate_fn(feat_extractor, lang_encoder, ds.sample_rate)

from torch.utils.data import DataLoader

train_loader = DataLoader(ds, batch_size=16, collate_fn=collate_fn, drop_last=True, shuffle=True)

from tqdm.auto import tqdm

## estimate batch count (all categories)
len(pcm_list) * 20 * 60 / 5 / 16

%%time 
for x in tqdm(train_loader):
    pass

import psutil
mem = psutil.Process().memory_info()
print("RSS(GB): ", mem.rss/1024**3)
print("Virtual(GB): ", mem.vms/1024**3)
