==> var/40.01-corpus-check.ipynb <==
# Abo16 Corpus

import pandas as pd

lang_map = pd.read_csv("../data/lang_map.csv", names="filename lang".split())
# lang_map = lang_map.loc[~lang_map.filename.str.contains("c5008-22"), :]

lang_map.lang.value_counts()

file_items = lang_map.iloc[::-1].reset_index(drop=True).to_dict(orient="index")

import imageio
from pathlib import Path
from matplotlib import pyplot as plt
from PIL import Image, ImageDraw, ImageFont
import cv2
import face_recognition
from tqdm.auto import tqdm

def make_video_path(video_name):
    video_dir = Path("../data/videos")
    date_dir = video_name.split("-")[1].replace(".mp4", "")
    video_path = video_dir / date_dir / video_name
    return video_path


def snapshot(video_path):    
    video = cv2.VideoCapture(str(video_path))
    fps = video.get(cv2.CAP_PROP_FPS)
    video.set(cv2.CAP_PROP_POS_FRAMES, int(23 * fps))
    ok, im = video.read()
    if not ok:
        raise IOError("Error when reading video")
    im = cv2.cvtColor(im, cv2.COLOR_BGR2RGB)
    return im

def annotate_image(im_arr):
    im = Image.fromarray(im_arr)
    font = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf", 36)
    draw = ImageDraw.Draw(im)        
    draw.rectangle([(40,10), (1000, 110)], fill="lightgray")
    draw.text((50,20), video_name, fill="blue", font=font)
    draw.text((50,60), lang, fill="blue", font=font)  
    faces = face_recognition.face_locations(im_arr)    
    for face_x in faces:
        t,r,b,l = face_x        
        draw.rectangle([(l,t),(r,b)], outline="orange", width=4)
    return im

sandbox_dir = Path("../data/sandbox/lang_checks")
sandbox_dir.mkdir(parents=True, exist_ok=True)

fout = open(sandbox_dir/"error_list.txt", "w")
for item_x in tqdm(file_items.values()):
    try:
        video_name = item_x["filename"]
        lang = item_x["lang"]
        video_path = make_video_path(video_name)
        im_arr = snapshot(video_path)
        im_annot = annotate_image(im_arr)
        im_annot.save(sandbox_dir/video_name.replace(".mp4", ".lang.jpg"))
    except Exception as ex:        
        fout.write(f"{video_name}, {ex}\n")
        print("Error", item_x)          
fout.close()
