==> var/40.02-corpus-selection.ipynb <==
# Selection - Abo16 Corpus

import pandas as pd

lang_map = pd.read_csv("../data/lang_map.csv", names="filename lang".split())
# lang_map = lang_map.loc[~lang_map.filename.str.contains("c5008-22"), :]

error_list = pd.read_csv("../data/sandbox/lang_checks/error_list.txt", names="filename note".split())
check_list = pd.read_csv("../data/sandbox/lang_checks/check_list.txt", names="filename note".split())
drop_list = pd.concat([error_list, check_list], axis=0)

lang_list = lang_map.loc[~lang_map.filename.isin(drop_list.filename)]
lang_list.lang.value_counts()

16*9

from collections import Counter
counter = Counter()
video_items = []
for _, row in lang_list.iloc[::-1].iterrows():    
    counter[row.lang] += 1
    if counter[row.lang] <= 9:
        video_items.append((row.filename, row.lang))

selected = pd.DataFrame.from_records(video_items, columns="filename lang".split())

selected.to_csv("../data/abo16_list.txt", index=None)
