==> var/40.20-extract-an-lang.ipynb <==
import pandas as pd
from pathlib import Path
import pydub
from tqdm.auto import tqdm

abo16_list = pd.read_csv("../data/abo16_list.txt")
anmp3_dir = Path("../data/anmp3")
anmp3_dir.mkdir(parents=True, exist_ok=True)

abo16_list.head()

def make_mp3_path(video_name):
    mp3_dir = Path("../data/mp3")    
    video_path = mp3_dir/ video_name.replace(".mp4", ".mp3")
    return video_path

mp3_dir = Path("../data/mp3")

import numpy as np

misc_index = np.arange(abo16_list.shape[0])
np.random.seed(123)
np.random.shuffle(misc_index,)
# misc_index = misc_index[:9]
misc_index = []

# thau_list = abo16_list.loc[abo16_list.lang=='Thau']
paiwan_list = abo16_list.loc[abo16_list.lang=='Paiwan']

for idx in tqdm(range(abo16_list.shape[0])):
# for idx in tqdm(range(paiwan_list.shape[0])):
    basename = abo16_list.iloc[idx].filename
    out_path = anmp3_dir / basename.replace(".mp4", ".mp3")    
        
    lang = abo16_list.iloc[idx].lang
    mp3_path = make_mp3_path(basename)
    mp3 = pydub.AudioSegment.from_mp3(mp3_path)
    anchor_path = "../data/sandbox/anchor_segs/" + basename.replace(".mp4", ".anchor.csv")
    anchor_seg = pd.read_csv(anchor_path, index_col=0)
    an_offsets = anchor_seg.loc[anchor_seg.is_anchor].offset.tolist()
    misc_offsets = anchor_seg.iloc[20:].loc[~anchor_seg.is_anchor].offset.tolist()
    if len(an_offsets) < 1:
        print("INFO: fewer than 10 anchor footage detected in ", basename)
    misc_offsets = misc_offsets[:len(an_offsets)]
    concat = pydub.AudioSegment.empty()
    
    if idx in misc_index:
        misc = pydub.AudioSegment.empty()
    else:
        misc = None
        
    cursor = 0
    for offset_i, offset_x in enumerate(an_offsets):    
        if offset_x < cursor: continue    
        if misc is not None:            
            misc_x = misc_offsets[offset_i]            
            misc += mp3[misc_x*1000: (misc_x+5)*1000]
        concat += mp3[offset_x*1000: (offset_x+5)*1000]
        cursor = offset_x+5
    
    out_path = anmp3_dir / basename.replace(".mp4", ".mp3")
    concat.export(out_path)
    if misc is not None:
        misc_path = anmp3_dir / ("misc_" + basename.replace(".mp4", ".mp3"))
        misc.export(misc_path)
    

## Anchor stat

anchor_map = {}
for idx in tqdm(range(abo16_list.shape[0])):
# for idx in tqdm(range(paiwan_list.shape[0])):
    basename = abo16_list.iloc[idx].filename
    out_path = anmp3_dir / basename.replace(".mp4", ".mp3")    
        
    lang = abo16_list.iloc[idx].lang        
    anchor_path = "../data/sandbox/anchor_segs/" + basename.replace(".mp4", ".anchor.csv")
    anchor_seg = pd.read_csv(anchor_path, index_col=0)
    anchor_name = anchor_seg.loc[anchor_seg.is_anchor].anchor_name
    try:
        counter = anchor_map.setdefault(lang, Counter())
        counter[anchor_name[0]] += 1
    except Exception:
        pass

{k: len(v) for k, v in anchor_map.items()}
