==> var/40.22-prepare-abo16-an.ipynb <==
import pandas as pd
from pathlib import Path
import pydub
from tqdm.auto import tqdm
import numpy as np

abo16_list = pd.read_csv("../data/abo16_list.txt")
anmp3_dir = Path("../data/anmp3")
anmp3_dir.mkdir(parents=True, exist_ok=True)

abo16_list.head()

abo16_map = abo16_list.to_dict(orient="records")
abo16_map = {x["filename"].replace(".mp4", ""): x["lang"] for x in abo16_map}

lang_map = {}
pcmlen_map = {}
for path_x in Path("../data/anpcm").glob("*.pcm"):
    lang = abo16_map.get(path_x.stem, "misc")
    pcm = np.memmap(path_x, np.dtype('int16'), 'r')
    pcmlen_map[path_x.stem+".mp3"] = pcm.shape[0]
    lang_map.setdefault(lang, []).append(path_x.stem+".mp3")

abo16_an_splits = {"train": [], "test": []}
for lang, fnames in lang_map.items():    
    abo16_an_splits["test"].append((fnames[0], lang, pcmlen_map[fnames[0]]))
    abo16_an_splits["train"].extend([(x, lang, pcmlen_map[x]) for x in fnames[1:]])

len(lang_map)

[(k, len(v)) for k, v in lang_map.items()]

import json
with open("../data/abo16_an_splits.json", "w") as fout:
    json.dump(abo16_an_splits, fout)

## LangMap

lang_map

with open("../data/abo16_lang_map.csv", "w") as fout:
    for k, fnames in lang_map.items():
        for fname_x in fnames:
            fout.write(f"{fname_x},{k}\n")
